Add a wrapper script around jhbuild
The jhbuild-moblin2 shell script is a simple wrapper that sets
the jhbuildrc file command line option and then passes the rest
of the options to jhbuild.

In order to make it work correctly, moblin2-jhbuild will now
install the jhbuildrc file in $HOME/.jhbuildrc-moblin2 and the
jhbuild-moblin2 script under $bindir (by default, $HOME/bin).

This should make setting up the whole machinery as easy as:

        $ git clone moblin:moblin2-jhbuild
        $ cd moblin2-jhbuild
        $ make install
        $ jhbuild-moblin2 build